Move pokemon forms to balance folder
