+++ src/data/balance/form-change-items.ts
@@ -1,0 +1,119 @@
+export enum FormChangeItem {
+  NONE,
+
+  ABOMASITE,
+  ABSOLITE,
+  AERODACTYLITE,
+  AGGRONITE,
+  ALAKAZITE,
+  ALTARIANITE,
+  AMPHAROSITE,
+  AUDINITE,
+  BANETTITE,
+  BEEDRILLITE,
+  BLASTOISINITE,
+  BLAZIKENITE,
+  CAMERUPTITE,
+  CHARIZARDITE_X,
+  CHARIZARDITE_Y,
+  DIANCITE,
+  GALLADITE,
+  GARCHOMPITE,
+  GARDEVOIRITE,
+  GENGARITE,
+  GLALITITE,
+  GYARADOSITE,
+  HERACRONITE,
+  HOUNDOOMINITE,
+  KANGASKHANITE,
+  LATIASITE,
+  LATIOSITE,
+  LOPUNNITE,
+  LUCARIONITE,
+  MANECTITE,
+  MAWILITE,
+  MEDICHAMITE,
+  METAGROSSITE,
+  MEWTWONITE_X,
+  MEWTWONITE_Y,
+  PIDGEOTITE,
+  PINSIRITE,
+  RAYQUAZITE,
+  SABLENITE,
+  SALAMENCITE,
+  SCEPTILITE,
+  SCIZORITE,
+  SHARPEDONITE,
+  SLOWBRONITE,
+  STEELIXITE,
+  SWAMPERTITE,
+  TYRANITARITE,
+  VENUSAURITE,
+
+  BLUE_ORB = 50,
+  RED_ORB,
+  ADAMANT_CRYSTAL,
+  LUSTROUS_GLOBE,
+  GRISEOUS_CORE,
+  REVEAL_GLASS,
+  MAX_MUSHROOMS,
+  DARK_STONE,
+  LIGHT_STONE,
+  PRISON_BOTTLE,
+  RUSTED_SWORD,
+  RUSTED_SHIELD,
+  ICY_REINS_OF_UNITY,
+  SHADOW_REINS_OF_UNITY,
+  ULTRANECROZIUM_Z,
+
+  SHARP_METEORITE = 100,
+  HARD_METEORITE,
+  SMOOTH_METEORITE,
+  GRACIDEA,
+  SHOCK_DRIVE,
+  BURN_DRIVE,
+  CHILL_DRIVE,
+  DOUSE_DRIVE,
+  N_SOLARIZER,
+  N_LUNARIZER,
+  WELLSPRING_MASK,
+  HEARTHFLAME_MASK,
+  CORNERSTONE_MASK,
+  FIST_PLATE,
+  SKY_PLATE,
+  TOXIC_PLATE,
+  EARTH_PLATE,
+  STONE_PLATE,
+  INSECT_PLATE,
+  SPOOKY_PLATE,
+  IRON_PLATE,
+  FLAME_PLATE,
+  SPLASH_PLATE,
+  MEADOW_PLATE,
+  ZAP_PLATE,
+  MIND_PLATE,
+  ICICLE_PLATE,
+  DRACO_PLATE,
+  DREAD_PLATE,
+  PIXIE_PLATE,
+  BLANK_PLATE, // TODO: Find a potential use for this
+  LEGEND_PLATE, // TODO: Find a potential use for this
+  FIGHTING_MEMORY,
+  FLYING_MEMORY,
+  POISON_MEMORY,
+  GROUND_MEMORY,
+  ROCK_MEMORY,
+  BUG_MEMORY,
+  GHOST_MEMORY,
+  STEEL_MEMORY,
+  FIRE_MEMORY,
+  WATER_MEMORY,
+  GRASS_MEMORY,
+  ELECTRIC_MEMORY,
+  PSYCHIC_MEMORY,
+  ICE_MEMORY,
+  DRAGON_MEMORY,
+  DARK_MEMORY,
+  FAIRY_MEMORY,
+  NORMAL_MEMORY // TODO: Find a potential use for this
+}
