+++ src/data/challenge.ts
@@ -15,7 +15,7 @@ import { Nature } from "#app/data/nature";
 import { Moves } from "#enums/moves";
 import { TypeColor, TypeShadow } from "#enums/color";
 import { pokemonEvolutions } from "#app/data/balance/pokemon-evolutions";
-import { pokemonFormChanges } from "#app/data/pokemon-forms";
+import { pokemonFormChanges } from "#app/data/balance/pokemon-forms";
 
 /** A constant for the default max cost of the starting party before a run */
 const DEFAULT_PARTY_MAX_COST = 10;
