+++ src/data/mystery-encounters/mystery-encounter-requirements.ts
@@ -7,7 +7,9 @@ import { Species } from "#enums/species";
 import { TimeOfDay } from "#enums/time-of-day";
 import { Nature } from "#app/data/nature";
 import { EvolutionItem, pokemonEvolutions } from "#app/data/balance/pokemon-evolutions";
-import { FormChangeItem, pokemonFormChanges, SpeciesFormChangeItemTrigger } from "#app/data/pokemon-forms";
+import { SpeciesFormChangeItemTrigger } from "#app/data/pokemon-forms";
+import { pokemonFormChanges } from "#app/data/balance/pokemon-forms";
+import { FormChangeItem } from "#app/data/balance/form-change-items";
 import { StatusEffect } from "#app/data/status-effect";
 import { Type } from "#app/data/type";
 import { WeatherType } from "#app/data/weather";
